+++ src/main/java/it/eng/parer/soft/delete/beans/IRegistrazioneRichiesteDao.java
@@ -1,18 +1,14 @@
 /*
  * Engineering Ingegneria Informatica S.p.A.
  *
- * Copyright (C) 2023 Regione Emilia-Romagna
- * <p/>
- * This program is free software: you can redistribute it and/or modify it under the terms of
- * the GNU Affero General Public License as published by the Free Software Foundation,
- * either version 3 of the License, or (at your option) any later version.
- * <p/>
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Affero General Public License for more details.
- * <p/>
- * You should have received a copy of the GNU Affero General Public License along with this program.
- * If not, see <https://www.gnu.org/licenses/>.
+ * Copyright (C) 2023 Regione Emilia-Romagna <p/> This program is free software: you can
+ * redistribute it and/or modify it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the License, or (at your option)
+ * any later version. <p/> This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ * PARTICULAR PURPOSE. See the GNU Affero General Public License for more details. <p/> You should
+ * have received a copy of the GNU Affero General Public License along with this program. If not,
+ * see <https://www.gnu.org/licenses/>.
  */
 
 package it.eng.parer.soft.delete.beans;
@@ -44,6 +40,12 @@ public interface IRegistrazioneRichiesteDao {
     boolean existAroRichRestArch(BigDecimal idStrut, BigDecimal idRichRestArch)
             throws AppGenericPersistenceException;
 
+    Long getIdRichScartoVersEvasa(BigDecimal idStrut, BigDecimal idRichiestaSacer)
+            throws AppGenericPersistenceException;
+
+    boolean existAroRichScartoVers(BigDecimal idStrut, BigDecimal idRichScartoVers)
+            throws AppGenericPersistenceException;
+
     AroRichSoftDelete getAroRichSoftDeleteContainingUd(Long idUnitaDoc, Long idRichSoftDelete)
             throws AppGenericPersistenceException;
 
@@ -52,6 +54,9 @@ public interface IRegistrazioneRichiesteDao {
 
     AroRichSoftDelete getAroRichSoftDeleteContainingRichRestArch(Long idRichRestArch,
             Long idRichSoftDelete, Long idStrut) throws AppGenericPersistenceException;
+
+    AroRichSoftDelete getAroRichSoftDeleteContainingRichScartoVers(Long idRichScartoVers,
+            Long idRichSoftDelete) throws AppGenericPersistenceException;
 
     List<DmUdDel> recuperaUnitaDocDaRichiesta(Long idStrut, BigDecimal idRichiestaSacer)
             throws AppGenericPersistenceException;
@@ -66,6 +71,8 @@ public interface IRegistrazioneRichiesteDao {
     boolean isUdNonAnnullata(long idUnitaDoc) throws AppGenericPersistenceException;
 
     boolean isUdNonRestituita(long idUnitaDoc) throws AppGenericPersistenceException;
+
+    boolean isUdNonScartata(long idUnitaDoc) throws AppGenericPersistenceException;
 
     List<AroErrRichSoftDelete> getAroErrRichSoftDeleteByGravity(long idItemRichSoftDelete,
             String tiGravita) throws AppGenericPersistenceException;