+++ src/main/java/it/eng/parer/soft/delete/beans/utils/CostantiDB.java
@@ -1,18 +1,14 @@
 /*
  * Engineering Ingegneria Informatica S.p.A.
  *
- * Copyright (C) 2023 Regione Emilia-Romagna
- * <p/>
- * This program is free software: you can redistribute it and/or modify it under the terms of
- * the GNU Affero General Public License as published by the Free Software Foundation,
- * either version 3 of the License, or (at your option) any later version.
- * <p/>
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Affero General Public License for more details.
- * <p/>
- * You should have received a copy of the GNU Affero General Public License along with this program.
- * If not, see <https://www.gnu.org/licenses/>.
+ * Copyright (C) 2023 Regione Emilia-Romagna <p/> This program is free software: you can
+ * redistribute it and/or modify it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the License, or (at your option)
+ * any later version. <p/> This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ * PARTICULAR PURPOSE. See the GNU Affero General Public License for more details. <p/> You should
+ * have received a copy of the GNU Affero General Public License along with this program. If not,
+ * see <https://www.gnu.org/licenses/>.
  */
 
 /*
@@ -181,7 +177,7 @@ public class CostantiDB {
 
         ITEM_GIA_PRESENTE, ITEM_IN_CORSO_DI_ELAB, ITEM_IN_CORSO_DI_ANNUL,
         ITEM_IN_CORSO_DI_REST_ARCH, ITEM_IN_CORSO_DI_SCARTO, ITEM_NON_ANNULLATO,
-        ITEM_NON_RESTITUITO, ITEM_NON_ESISTE, STATO_CONSERV_NON_AMMESSO;
+        ITEM_NON_RESTITUITO, ITEM_NON_SCARTATO, ITEM_NON_ESISTE, STATO_CONSERV_NON_AMMESSO;
 
         public static String[] getStatiControlloItem() {
             // <<<<<<< Ritorna tutti gli stati tranne ITEM_NON_ESISTE e