+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.3.0 (31-03-2026)
+
+### Novità: 1
+- [#39660](https://parermine.regione.emilia-romagna.it/issues/39660) Estensione cancellazione logica (Soft Delete) per supporto Scarto
+
 ## 1.2.0 (03-02-2026)
 
 ### Novità: 1