+++ RELEASE-NOTES.md
@@ -1,4 +1,4 @@
-## 1.2.0 (03-02-2026)
+## 1.3.0 (31-03-2026)
 
 ### Novità: 1
-- [#39105](https://parermine.regione.emilia-romagna.it/issues/39105) Estensione cancellazione logica (Soft Delete) per supporto restituzione archivio
+- [#39660](https://parermine.regione.emilia-romagna.it/issues/39660) Estensione cancellazione logica (Soft Delete) per supporto Scarto